Add OsPlacesApi::Client and 'locations_for_postcode' method
This method searches for locations based on postcode. See
https://osdatahub.os.uk/docs/places/technicalSpecification
for the technical specification.

It uses the OAuth approach to authenticating the request - see
https://osdatahub.os.uk/docs/oauth2/gettingStarted.

Note that the API uses the "DPA" dataset by default, so does not
need to be specified in the request. In a later commit, we could
extend the method to allow specifying extra options in the request,
such as overriding the dataset choice. Until then,
[YAGNI](https://martinfowler.com/bliki/Yagni.html).

It's worth noting that the AccessTokenManager class is separate
because:

1) We don't want to expose the access_token method on the client
   itself
2) Access tokens are valid for a short period of time, so we can
   reuse the same access token until it expires. We want to
   encapsulate any business logic associated with that, rather
   than overcomplicate the client.
   That said, we won't bother adding that logic until we need it
   (again, YAGNI).