Raise InvalidPostcodeProvided if we send a bad request to OS API
This situation should not happen, as we intend to validate
postcodes using the PostcodeValidator _before_ any requests are
made to the OS Places API. However, we should handle this
situation if it occurs.

Integrating the validation will happen later, at the model /
controller level. When Locations API gets a request from a frontend
app, it will look for an existing Postcode record in the database
(implying we have a cached result already), and if there isn't one,
it will attempt to create a record, using the postcode validator
to validate the postcode. Only then will a request be made to OS
Places API via our OsPlacesApi::Client.