Add 'Postcode' model
Adds model to be saved to the database, acting as a caching layer
between Locations API and OS Places API.

The PostcodeValidator specs allow all sorts of extra characters
in the postcode, such as '(', ')', '+' and '-'. We don't want to
store these additional characters, as they're only used for
formatting. I've exposed the characters in a new 'DISALLOWED_CHARS'
regex so that the regex can be shared between the validator and
the normaliser.

The 'removes disallowed characters from the provided postcode'
test isn't a full list of all the disallowed characters, but
is based on the list in postcode_validator_spec.rb.

Trello: https://trello.com/c/olgAVxu0/2778-implement-caching-database-for-locations-api-5