Retrieve from database before fetching from OS Places API
Now that we have a database table to store OS Places API results
to, we should return results from our database cache rather than
make requests to OS Places API, if we've already cached the
response for the given postcode.

I've rearchitected the client_spec.rb file a little to account for
this new behaviour. All the API response validation should now
live inside the "postcode doesn't exist in the database" context,
as we now only make requests to OS Places API when it's missing
from the database.

I've added a new test for when the postcode exists in the database.
As part of this, I've had to override the equality operator in
the Location model, as otherwise the test would fail (saying that
the Location instances aren't equal) when in fact these are only
data models and have identical data, so in other words are equal.