Cache results in database after making request to OS Places API
I did consider writing a test that only 1 record gets created
even after multiple calls to `locations_for_postcode`, but this is
already covered in postcode_spec.rb. I also considered a test to
ensure that only 1 network request is made, but this is covered
by the 'the postcode exists in the database' test in client_spec.rb.