Prevent `locations_for_postcode` race condition
As pointed out in code review:
https://github.com/alphagov/locations-api/pull/28#discussion_r785915100

If two users request the same postcode at the same time, and the
postcode isn't in the database yet, then there's a risk that User
A's request will cause the postcode to be created in the database
after User B's request fails to find the record in the database,
meaning User B's request would ultimately make a `create!` call
that would raise an exception due to the duplication.

The workaround for the race condition is a simple `unless`, though
the test is a bit gnarly.