Limit to one worker at a time
We initialise a `PostcodesCollectionWorker` on startup of the Rails
app, but because we deploy Locations API to multiple instances,
that means we spin up multiple `PostcodesCollectionWorker` workers.
Taking inspiration from Email Alert API ([1]), we can use the
`with_advisory_lock` gem to limit Sidekiq to just one worker.

I've had to move the iteration back to a standard `loop` syntax,
rather than re-instantiate the worker, otherwise Sidekiq just
hangs on the original worker.

I've also had to use the `with_advisory_lock` trick within the
`PostcodesCollectionWorker` itself, otherwise it seemed to spawn
a larger number of `ProcessPostcodeWorker`s than it should have.
It's not an easy thing to test, but we do need some sort of
safeguard to stop the code being accidentally deleted, so hopefully
this imperative style test documents the reasons around the code
a bit.

Finally, I had to use the `Postcode.uncached` otherwise the worker
seemed to cache the initial result of `Postcode.pluck`, meaning
that any postcodes added to the database whilst the original
`PostcodesCollectionWorker` was running would fail to be updated.

[1]: https://github.com/alphagov/email-alert-api/blob/3c477520682a6eb1ac1fcd1d945ab2a652efd199/app/workers/daily_digest_initiator_worker.rb#L3-L5