Sort postcodes by `updated_at` before updating them
When we have millions of postcodes, it will take days to process
each postcode. If we do a deployment in the meantime, the worker
queue is cleared, and we'd be starting again at the beginning.

By sorting the postcodes by `updated_at` first, we ensure we'll
always be updating the oldest postcodes first.