Bump concurrency to 10
We had originally planned to have one long-running PostcodesCollectionWorker,
and process one postcode worker at a time (spawned from the above).
But we actually now want to process 3 postcodes at a time (see
previous commit).

Whilst we could set a concurrency of 3 to enforce this, that does
mean changing the value in multiple places if we were ever to change
this number in the future. By setting it to 10, we've introduced
a safeguard of sorts (OS Places API has a rate limit of 600 per
minute, i.e. 10 per second), and our use of `ApplicationRecord`'s
`with_advisory_lock` should prevent us ever spawning more than
one active `PostcodesCollectionWorker` at a time, which then
itself spawns the number of workers corresponding to
`POSTCODES_PER_SECOND` (currently `3`).

So even with a concurrency set to 10, we shouldn't expect to see
more than around 4 workers at any one time, because of how we've
architected the workers themselves. We can easily boost the
concurrency numbers by changing the `POSTCODES_PER_SECOND` in
future (provided it is less than 10).