Use sidekiq-uniq-jobs and sidekiq-scheduler for predictability
sidekiq-uniq-jobs prevents duplicate jobs. This is useful, as
Sidekiq runs on each of the EC2 instances Locations API runs on,
meaning it is difficult to limit instantiation of workers to
predictable levels.

We have therefore removed the 'with_advisory_lock' dependency,
which was the previous attempt at limiting spawned workers to
one. It never seemed to work very well.

Having PostcodesCollectionWorker spawn on upstart, and being
responsible for spawning the next PostcodesCollectionWorker in
the sequence, was brittle. If anything prevented PostcodesCollectionWorker
from spawning the next PostcodesCollectionWorker, the sequence
would simply die.

We now use sidekiq-scheduler to spawn a PostcodesCollectionWorker
every second - which is what we were trying to do by hand before.
This is much less brittle, and much more accurate. Deploying
to Integration, we can see exactly 300 PostcodesCollectionWorker
workers are instantiated over 5 minutes, leading to exactly
900 ProcessPostcodeWorker workers.