Apply postcode normalisation consistently around Locations API
Previously, we were only normalising postcodes at the point of
committing them to the database. We weren't doing any such
normalisation at the OsPlacesApi::Client level, meaning that if
a user passed a postcode that already existed in the database,
but was in a slightly different format (e.g. lowercase, or space
added), OsPlacesApi::Client would treat it as a new postcode and
perform a request.

As well as performing the unnecessary request, the application
would then raise a validation error when attempting to commit the
result to the database, as at this point the postcode would be
normalised and Locations API would realise that the postcode is
not unique, thus PostcodeValidator would raise an exception.

We now normalise the postcode at the point it enters Locations API,
so these kinds of inconsistencies should stop happening.

Fixes #70