Install `sentry-sidekiq` gem and remove manual call to GovukError
In #74, we added a manual call to `GovukError.notify` when an
exception was raised in Sidekiq. This was because Sentry would
not receive any errors from Sidekiq.

However, govuk_app_config does support Sidekiq Sentry error
reporting - it just needs the `sentry-sidekiq` gem adding.
This is explicitly warned about as of version 4.6.0:
https://github.com/alphagov/govuk_app_config/blob/main/CHANGELOG.md#460

By adding this dependency, we should be able to remove the
duplicate config.