Make "select oldest 3 postcode rows" more efficient
Now that we have over 2 million rows in the database, the old
`sort_by` method was taking a long time to complete - in fact, it
was kicking me out of the Rails console. Using the `.order` method
is far, far quicker.

Also swapping the order of the `.first` and the `.pluck` led to
significant performance improvements too.