Force `updated_at` to update even if no changes to the DB record
If we have a postcode's results cached in the database, and if
the response from OS Places API is identical the next time we
attempt to update the postcode (which we expect to be the case the
majority of the time), no `UPDATE` SQL query actually gets
executed, as there is nothing to update.

However, this means the `updated_at` property remains unchanged,
and as we use this property to decide which postcode to attempt
to update next ([1]), Locations API gets stuck attempting to
update the same few postcodes on a loop.

In this commit, we change Locations API to update the `updated_at`
field even if there has been no change to the contents of the
record in the database. We can therefore accurately use this
field to figure out when the last attempt was made to 'refresh'
the postcode via OS Places API.

This should fix the bug whereby no postcodes are being updated
anymore because the 3 oldest postcodes' results are unchanged,
and thus the 3 oldest postcodes would continue to be the 'oldest'
postcodes.

[1]: https://github.com/alphagov/locations-api/blob/a0605a6608f5d6d2c8f8de37ffb4d0d7f88dee33/app/workers/postcodes_collection_worker.rb#L9