Return 400 Bad Request when invalid postcode provided
We weren't handling the `OsPlacesApi::InvalidPostcodeProvided`
exception anywhere, meaning these requests would return a 500
error to GDS API Adapters ("Internal Server Error").
Note that GDS API Adapters has a test that invalid postcodes
return a 400 response, but this only applies to inputs that
don't match the postcode regex; postcodes that match the regex
but aren't valid postcodes return a 500. See
https://github.com/alphagov/gds-api-adapters/blob/809c007610f90e65ee6078c1ff2e6b57ae188546/test/locations_api_test.rb#L91-L93

Instead we should return 400, which indicates a client error,
which is what this is. The system can't be expected to return a
result for a postcode that is invalid.

Given that this is a client error, there is nothing for us to fix,
so any Sentry logging would just be noise. We therefore decide not
to log these responses in Sentry. See
https://docs.publishing.service.gov.uk/manual/sentry.html#deciding-whether-to-log-errors-in-sentry

We also decide not to log the bad request locally using
`Rails.logger.warn`. This information is already being logged in
Logit/Kibana, so we can extract bad inputs for analysis later if
needed.

Trello: https://trello.com/c/6vdbP8Kz/2986-fix-remove-locations-api-sentry-errors