Add GitHub Action workflow for verifying pact tests
This moves the pact verification to a GitHub Action and will no longer
be performed by a Jenkins build. This workflow will also be called by
GDS API Adapters as part of it's CI build to assert that newly generated
pacts are valid against this app. To avoid running this same task twice
`pact:verify` is no longer part of the Jenkins test actions and instead
the other rake default steps are included.

For more details see https://github.com/alphagov/gds-api-adapters/pull/1175