Allow Kircaldy postcodes
KY was incorrectly classified as a non-geographic postcode,
in fact KY covers Kircaldy, and it's specifically KYx-
postcodes we want to avoid (Caymans Islands postcodes). The
postcode normalisation process in the validator makes it
impossible to spot these, so we split them out and check
against the non-normalised postcode. This has a slight cost
of having to call PostcodeHelper.normalise twice in a validation,
but it's so lightweight that that shouldn't affect us.