Allow non-filtered results to contribute to average lat/lon
Sometimes we get a result set (example: TR21 0LW) where all the results
are filtered out as not suitable for displaying, but which still give us
a lat/lon suitable for geocoding. This change calculates lat/lon before
filtering, so we can now return results for this situation (av lat/lon
with no individual results) rather than an error.

We remove the test that an error gets raised if we have no results after
filtering, because this is now a valid thing to return.